# docs/swagger.yaml
basePath: /api
definitions:
...
  ds.LoginRequest:
    description: Данные для входа в систему
    properties:
      login:
        example: doctor_user
        type: string
      password:
        example: securepassword123
        type: string
    required:
    - login
    - password
    type: object
  ds.RegisterRequest:
    description: Данные для регистрации нового пользователя
    properties:
      login:
        example: new_user
        type: string
      password:
        example: securepassword123
        type: string
      role:
        enum:
        - patient
        - doctor
        - admin
        example: patient
        type: string
    required:
    - login
    - password
    type: object
...
  ds.UpdateUserRequest:
    description: Данные для обновления профиля пользователя
    properties:
      login:
        example: new_login
        type: string
      role:
        enum:
        - patient
        - doctor
        - admin
        example: doctor
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@nutriscan.ru
    name: API Support
    url: http://localhost:8081
  description: API для расчета индекса нутритивной недостаточности
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: NutriScan API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Выполняет вход пользователя в систему и возвращает JWT токен
      parameters:
      - description: Данные для входа
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/ds.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Неверные учетные данные
          schema:
            additionalProperties: true
            type: object
      summary: Аутентификация пользователя
      tags:
      - Аутентификация
  /auth/logout:
    post:
      description: Выполняет выход пользователя и добавляет токен в blacklist
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный токен
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выход из системы
      tags:
      - Аутентификация
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные для регистрации
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/ds.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Пользователь уже существует
          schema:
            additionalProperties: true
            type: object
      summary: Регистрация пользователя
      tags:
      - Аутентификация
...
  /research_biomarkers/{research_id}/biomarkers/{biomarker_id}:
    delete:
      description: Удаляет биомаркер из исследования (только для врачей)
      parameters:
      - description: ID исследования
        in: path
        name: research_id
        required: true
        type: integer
      - description: ID биомаркера
        in: path
        name: biomarker_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Биомаркер удален из исследования
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Исследование или биомаркер не найдены
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить биомаркер из исследования
      tags:
      - Исследования
    put:
      consumes:
      - application/json
      description: Обновляет значение пациента для биомаркера в исследовании (только
        для врачей)
      parameters:
      - description: ID исследования
        in: path
        name: research_id
        required: true
        type: integer
      - description: ID биомаркера
        in: path
        name: biomarker_id
        required: true
        type: integer
      - description: Обновляемые поля
        in: body
        name: updates
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Значение обновлено
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Исследование или биомаркер не найдены
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить значение биомаркера в исследовании
      tags:
      - Исследования
...
  /users/profile:
    get:
      description: Возвращает информацию о пользователе по ID
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить профиль пользователя
      tags:
      - Пользователи
    put:
      consumes:
      - application/json
      description: Обновляет информацию о пользователе
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      - description: Обновляемые поля
        in: body
        name: updates
        required: true
        schema:
          $ref: '#/definitions/ds.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль обновлен
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить профиль пользователя
      tags:
      - Пользователи
securityDefinitions:
  ApiKeyAuth:
    description: Введите "Bearer [ваш_JWT_токен]" для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...